Add files via upload
This Handwritten Digit Recognition using Logistic Regression" repository is a machine learning project that aims to recognize handwritten digits using logistic regression as the classification algorithm. The project uses the MNIST dataset, which is a popular benchmark dataset for image classification. The notebook also includes visualizations that show the digit images and their corresponding labels.